выполнен рефакторинг
