+++ js/api.js
@@ -1,7 +1,7 @@
 const BASE_URL = 'https://28.javascript.pages.academy/kekstagram';
 const Route = {
   GET_DATA: '/data',
-  SEND_DATA: '/404',
+  SEND_DATA: '/',
 };
 
 const Method = {
@@ -9,12 +9,9 @@ const Method = {
   POST: 'POST',
 };
 
-const ErrorText = {
-  GET_DATA: 'Не удалось загрузить данные. Попробуйте обновить страницу',
-  SEND_DATA: 'Не удалось отправить форму. Попробуйте ещё раз',
-};
+const ERROR_TEXT = 'Не удалось загрузить данные. Попробуйте обновить страницу';
 
-const load = (route, errorText, method = Method.GET, body = null) =>
+const load = (route, errorText = ERROR_TEXT, method = Method.GET, body = null) =>
   fetch(`${BASE_URL}${route}`, {method, body})
     .then((response) => {
       if (!response.ok) {
@@ -26,9 +23,9 @@ const load = (route, errorText, method = Method.GET, body = null) =>
       throw new Error(errorText);
     });
 
-const getData = () => load(Route.GET_DATA, ErrorText.GET_DATA);
+const getData = () => load(Route.GET_DATA);
 
-const sendData = (body) => load(Route.SEND_DATA, ErrorText.SEND_DATA, Method.POST, body);
+const sendData = (body) => load(Route.SEND_DATA, undefined, Method.POST, body);
 
 
 export {getData, sendData};
